% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcombination.R
\name{inverse.logit}
\alias{inverse.logit}
\title{Sigmoid (inverse logit) function}
\usage{
inverse.logit(x)
}
\arguments{
\item{x}{numeric vector or scalar}
}
\value{
numeric vector or scalar of same length as x
}
\description{
Returns the sigmoid transformed value(s) of x
}
\note{
Nothing to be noted
}
\examples{
inverse.logit(-3.4)
}
\author{
Jan Brink Valentin
}
