% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcombination.R
\name{is.formula}
\alias{is.formula}
\title{Is object formula}
\usage{
is.formula(x)
}
\arguments{
\item{x}{any object}
}
\value{
logical
}
\description{
Returns logical indicating if x is a formula starting with "~"
}
\note{
Nothing to be noted
}
\examples{
is.formula(~x1+x2)
}
\author{
Jan Brink Valentin
}
