% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcombination.R
\name{inverse.logit}
\alias{inverse.logit}
\title{Inverse logit (sigmoid) function}
\usage{
inverse.logit(x)
}
\arguments{
\item{x}{numeric vector or scalar}
}
\value{
numeric vector or scalar of same length as x
}
\description{
Inverse logit (sigmoid) function
}
\examples{
inverse.logit(-3.4)
}
