% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcombination.R
\name{is.formula}
\alias{is.formula}
\title{Is object formula}
\usage{
is.formula(x)
}
\arguments{
\item{x}{any object}
}
\value{
boolean
}
\description{
Is object formula
}
\examples{
is.formula(~x1+x2)
}
