% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcombination.R
\name{nlcombination}
\alias{nlcombination}
\title{Non-linear combinations}
\usage{
nlcombination(
  x,
  coefs,
  vcov,
  exp.form = FALSE,
  inv.logit.form = FALSE,
  conf.level = 0.95
)
}
\description{
Non-linear combinations
}
